make worker for configuration loading lighter
